Draft implementations of parquet-cpp allocator and read-only file interfaces
